refactor actions names
