+++ bot/src/telegram/actions/assign-guard.js
