+++ bot/src/telegram/actions/delete-guard.js
