+++ bot/src/telegram/actions/get-guard-information.js
