+++ bot/src/telegram/actions/get-not-covered-guards.js
@@ -1,5 +1,5 @@
 import agenda from "../../init.js";
-import { GET_AVAILABLE_USER_FOR_ASSIGN } from "./get-available-user-for-assign-act.js";
+import { GET_AVAILABLE_USER_FOR_ASSIGN } from "./get-available-user-for-assign.js";
 import { newValueButton, newActionButton } from "../keyboard-utils.js";
 
 export const GET_NOT_COVERED_GUARDS = {
